requirements.txt
google-cloud-firestore==2.18.0
google-cloud-storage==2.10.0
requests==2.31.0 